+++ fmn/rules/__init__.py
@@ -3,11 +3,14 @@ from fmn.rules.askbot import *
 from fmn.rules.bodhi import *
 from fmn.rules.buildsys import *
 from fmn.rules.compose import *
+from fmn.rules.coprs import *
 from fmn.rules.fas import *
 from fmn.rules.fedbadges import *
+from fmn.rules.fedocal import *
 from fmn.rules.fedoratagger import *
 from fmn.rules.generic import *
 from fmn.rules.git import *
+from fmn.rules.logger import *
 from fmn.rules.mailman import *
 from fmn.rules.meetbot import *
 from fmn.rules.pkgdb import *