+++ fmn/rules/generic.py
@@ -19,13 +19,13 @@ def user_filter(config, message, fasnick=None, *args, **kw):
 def user_package_filter(config, message, fasnick=None, *args, **kw):
     """ All messages concerning user's packages
 
-    This rule rules out messages that related to packages where the
+    This rule includes messages that relate to packages where the
     specified user has **commit** ACLs.
     """
 
     fasnick = kw.get('fasnick', fasnick)
     if fasnick:
-        packages = fmn.rules.utils.get_packages_of_user(fasnick)
+        packages = fmn.rules.utils.get_packages_of_user(config, fasnick)
         return packages.intersection(fedmsg.meta.msg2packages(message))